Refactor and cleanup other login routes, add github oauth routes to main app
