+++ server/nlp_ssa/api/app.py
@@ -10,7 +10,8 @@ from api.routes.api.analysis_views import router as api_analysis_views
 from api.routes.api.article_data import router as api_article_data
 from api.routes.api.stocks import router as api_stocks
 from api.routes.api.users import router as api_users
-from api.routes.login import router as login
+from api.routes.auth.github import router as github_oauth
+from api.routes.auth.login import router as login
 from config import env_config
 
 
@@ -52,3 +53,4 @@ app.include_router(api_article_data.router)
 app.include_router(api_stocks.router)
 app.include_router(api_users.router)
 app.include_router(login.router, prefix="/api/auth")
+app.include_router(github_oauth.router, prefix="/api/auth")
