+++ server/nlp_ssa/api/routes/auth/login/router.py
@@ -1,9 +1,6 @@
 import logging
-import requests
 import secrets
-from urllib import parse
 from fastapi import APIRouter, Depends, Request, HTTPException, status
-from fastapi.responses import RedirectResponse
 from fastapi.security import OAuth2AuthorizationCodeBearer
 from fastapi_csrf_protect import CsrfProtect
 
@@ -11,7 +8,7 @@ from fastapi_csrf_protect import CsrfProtect
 from oauthlib.oauth2 import WebApplicationClient
 from typing import Annotated
 
-from api.routes.login.models import LoginResponse
+from api.routes.auth.login.models import LoginResponse
 from config import env_vars
 from db import db_session
 from models.user import User, UserFacade
@@ -20,57 +17,13 @@ from models.user import User, UserFacade
 logger = logging.getLogger(__name__)
 router = APIRouter()
 
-from rich import inspect
-
-GITHUB_OAUTH_URL = "https://github.com/login/oauth/authorize"
-GITHUB_OAUTH_TOKEN_URL = "https://github.com/login/oauth/access_token"
 
 oauth2_auth_code_scheme = OAuth2AuthorizationCodeBearer(
-    authorizationUrl=GITHUB_OAUTH_URL,
-    tokenUrl=GITHUB_OAUTH_TOKEN_URL,
+    authorizationUrl=env_vars.GITHUB_OAUTH_AUTH_URL,
+    tokenUrl=env_vars.GITHUB_OAUTH_TOKEN_URL,
     # scopes=["read:user"],
 )
 github_client = WebApplicationClient(env_vars.GITHUB_OAUTH_CLIENT_ID)
-
-
-def exchange_code(code):
-    params = {
-        "client_id": env_vars.GITHUB_OAUTH_CLIENT_ID,
-        "client_secret": env_vars.GITHUB_OAUTH_CLIENT_SECRET,
-        "code": code,
-    }
-
-    result = requests.post(
-        GITHUB_OAUTH_TOKEN_URL + parse.urlencode(params),
-        headers={"Accept": "application/json"},
-    )
-
-    inspect(result, methods=True, sort=True)
-
-    try:
-        response = result.json()
-    except Exception as e:
-        logger.error(e)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Nope :["
-        )
-
-    return response
-
-
-@router.get("/github-callback")
-async def github_oauth_callback(
-    fast_api_request: Request = None,
-    # token: str = Depends(oauth2_auth_code_scheme),
-):
-    inspect(fast_api_request, sort=True)
-    inspect(fast_api_request.query_params, methods=True, sort=True)
-
-    token_data = exchange_code(fast_api_request.query_params.get("code"))
-
-    inspect(token_data)
-
-    return RedirectResponse("https://nlp-ssa.dev/app")
 
 
 @router.get("/login", response_model=LoginResponse)
@@ -83,7 +36,7 @@ async def read_login(
     fast_api_request.state.github_oauth_state = secrets.token_urlsafe(16)
 
     github_url = github_client.prepare_request_uri(
-        GITHUB_OAUTH_URL,
+        env_vars.GITHUB_OAUTH_AUTH_URL,
         redirect_uri="https://nlp-ssa.dev/api/auth/github-callback",
         scope=["read:user"],
         state=fast_api_request.state.github_oauth_state,
